requirements.txt
torch>=2.1
torchvision>=0.16
numpy>=1.24
pillow>=10.0
omegaconf>=2.3
tensorboard>=2.15
pyiqa>=0.1.10
tqdm>=4.66
pytest>=7.4
